# use internal buffer's Buf impl for codec buffers